requirements.txt
streamlit==1.31.0
anthropic==0.18.1
chromadb==0.4.22
PyMuPDF>=1.24.0
sentence-transformers==2.3.1
python-dotenv==1.0.1
Pillow==10.2.0
